Added picard module
